requirements/prod.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
celery==5.3.1
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==3.2.16
django-cleanup==8.0.0
django-environ==0.9.0
django-redis==5.3.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
kombu==5.3.1
Pillow==9.4.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
python-dateutil==2.8.2
pytz==2023.3
redis==4.6.0
six==1.16.0
sorl-thumbnail==12.9.0
sqlparse==0.4.4
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6